Fix camera laggy when touching bound, add right bound and below bound for camera

+++ 03-Keyboard-States/Brick.cpp
@@ -3,8 +3,8 @@
 void CBrick::Render()
 {
 	CAnimations* animations = CAnimations::GetInstance();
-	int drawx = camera->TransitionX(x);
-	int drawy = camera->TransitionY(y);
+	float drawx = camera->TransitionX(x);
+	float drawy = camera->TransitionY(y);
 	if (((drawx < -30) || (drawx > 320)) &&
 		((drawy < -50) || (drawy > 240)))
 	{
