+++ 03-Keyboard-States/CLadder.cpp
@@ -19,8 +19,8 @@ void CLadder::LoadResource()
 void CLadder::Render()
 {
 	CAnimations* animations = CAnimations::GetInstance();
-	int drawx = camera->TransitionX(x);
-	int drawy = camera->TransitionY(y);
+	float drawx = camera->TransitionX(x);
+	float drawy = camera->TransitionY(y);
 	if (((drawx < -30) || (drawx > 320)) &&
 		((drawy < -50) || (drawy > 240)))
 	{
